FlowStack.swift.new
I didn't figure out how to make a proper pull request: the current <FlowStack.swift> version has the issue that it won't compile because the compiler can't infer the types in acceptable time. So the code was split up a little and some private variables and functions were added. This code should replace the former FlowStack.swift.